tests/requirements.txt
# Testing dependencies for RA documentation skills

# Core testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Coverage reporting
coverage>=7.3.0

# Word document manipulation
python-docx>=0.8.11

# Code quality
pytest-clarity>=1.0.1
pytest-sugar>=0.9.7

# Optional: For generating test reports
pytest-html>=3.2.0
pytest-json-report>=1.5.0
